dac: change max_voltage from int to float
in Ethernet, voltage levels are +- 2.5 V

+++ src/phyether/dac.py
@@ -11,7 +11,7 @@ class DAC:
     """
 
     def __init__(self, rise_time: float, on_time: float,
-                 high_symbol: int, symbol_step: int = 1, max_voltage: int = 2) -> None:
+                 high_symbol: int, symbol_step: int = 1, max_voltage: float = 2) -> None:
         """Convert symbols from digital to analog (piecewise linear)
 
         Formula for symbol to voltage conversion:
@@ -21,7 +21,7 @@ class DAC:
             rise_time (float): time to change from one voltage to another in nanoseconds
             on_time (float): duration of one symbol in nanoseconds
             high_symbol (int): highest symbol value in data
-            max_voltage (int): voltage for max_symbol
+            max_voltage (float): voltage for max_symbol
             symbol_step (int): difference between adjacent symbols
         """
         self.rise_time: UnitValue = u_ns(rise_time)
